feat: assistant use default models when no targets feat: display assistant fallback correctly feat: refactor thread header
